server/requirements.txt
Flask==1.0.2
IMDbPY==6.6